+++ src/main/kotlin/io/hirasawa/server/webserver/internalroutes/TestRoute.kt
@@ -1,6 +1,6 @@
-package io.hirasawa.server.webserver.routes
+package io.hirasawa.server.webserver.internalroutes
 
-import io.hirasawa.server.webserver.Route
+import io.hirasawa.server.webserver.route.Route
 import io.hirasawa.server.webserver.objects.Request
 import io.hirasawa.server.webserver.objects.Response
