+++ src/main/kotlin/io/hirasawa/server/webserver/route/Route.kt
@@ -1,4 +1,4 @@
-package io.hirasawa.server.webserver
+package io.hirasawa.server.webserver.route
 
 import io.hirasawa.server.webserver.objects.Request
 import io.hirasawa.server.webserver.objects.Response
